@@ -1,25 +1,14 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders, HttpResponse } from '@angular/common/http';
 
-import { Observable } from 'rxjs/Observable';
-import { of } from 'rxjs/observable/of';
-import { catchError, map, tap } from 'rxjs/operators';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/merge';
-import 'rxjs/add/operator/mergeMap';
-import 'rxjs/add/operator/isEmpty';
-import 'rxjs/add/operator/toArray';
-import 'rxjs/add/operator/publishReplay';
-
-import 'rxjs/add/observable/zip'
-import 'rxjs/add/observable/merge'
+import { Observable , of, zip, merge } from 'rxjs';
+import { catchError, map, tap , retry, toArray, flatMap, publishReplay, refCount } from 'rxjs/operators';
 
 import { MessageService } from './message.service';
 import { environment } from '../../environments/environment';
 import { ContentPage } from '../model/json/contentPage';
 import { Content } from '../model/json/content';
 import { Post } from '../model/post';
-import { retry } from 'rxjs/operators/retry';
 import { User } from '../model/json/user';
 
 @Injectable()
@@ -35,8 +24,9 @@ export class DataService {
     return this.httpClient.get<ContentPage>(pageContentUrl)
       .pipe(
           tap(r => this.log('fetched page ' + pageId)),
-          catchError(this.handleError('getPage ' + pageId, { content: { rendered: 'Page content not found' } })))
-      .map(v => v.content.rendered);
+          catchError(this.handleError('getPage ' + pageId, { content: { rendered: 'Page content not found' } })),
+          map(v => v.content.rendered)
+        );
   }
 
   checkNextPageExists(currentPage: number): Observable<boolean> {
@@ -46,8 +36,9 @@ export class DataService {
     return this.httpClient.get<ContentPage[]>(postsUrl)
       .pipe(
           tap(r => this.log('peeked for post ' + peekPost)),
-          catchError(this.handleError(request, [])))
-      .map(r => r.length > 0);
+          catchError(this.handleError(request, [])),
+          map(r => r.length > 0)
+        );
   }
 
   getPosts(page: number): Observable<Post[]> {
@@ -55,16 +46,18 @@ export class DataService {
     return this.httpClient.get<ContentPage[]>(postsUrl)
     .pipe(
         tap(r => this.log(`fetched posts from ${page}`)),
-        catchError(this.handleError<ContentPage[]>(postsUrl, [])))
-    .flatMap(cp => {
-        let ids = Array.from(new Set(cp.map(c => c.author)));
-        let map = new Map<number, Observable<string>>();
-        ids.forEach(v => map.set(v, this.getAuthor(v)));
-
-        let authors = Observable.merge(...cp.map<Observable<string>>(c => map.get(c.author)));
-        return Observable.zip(authors, cp, (a, c) => this.createPost(c, a));
-      }).toArray();
-  }
+        catchError(this.handleError<ContentPage[]>(postsUrl, [])),
+        flatMap(cp => {
+          let ids = Array.from(new Set(cp.map(c => c.author)));
+          let map = new Map<number, Observable<string>>();
+          ids.forEach(v => map.set(v, this.getAuthor(v)));
+  
+          let authors = merge(...cp.map<Observable<string>>(c => map.get(c.author)));
+          return zip(authors, cp, (a, c) => this.createPost(c, a));
+        }),
+        toArray()
+      );
+    }
 
   createPost(cp: ContentPage, author: string) : Post {
     return new Post(cp.content.rendered, cp.title.rendered, new Date(cp.date), author);
@@ -75,10 +68,11 @@ export class DataService {
     return this.httpClient.get<User>(authorUrl)
     .pipe(
       tap(r => this.log(`fetched username for user ${authorId}`)),
-      catchError(this.handleError<User>(authorUrl, new User())))
-    .map(u => u.name)
-    .publishReplay(1)
-    .refCount();
+      catchError(this.handleError<User>(authorUrl, new User())),
+      map(u => u.name),
+      publishReplay(1),
+      refCount()
+    );
   }
 
   /**
